// src/app/interfaces/mapdata.interface.ts
export interface IMapData {
  branch_id?: string;
  ordering?: string;
  country_id?: string;
  region_id?: string;
  area_id?: string;
  is_branch_country_region_area?: string;
  retailer_name?: string;
  location_name?: string;
  phone_no?: string;
  fax_no?: string;
  address?: string;
  latitude?: string;
  longitude?: string;
}